Broke down existing IME tests to multiple files to make it easier to add new tests for editable non-text nodes, like a list of attachments
